fixed some user-defined input issues

+++ parflow/python/test/clear.py
@@ -3,8 +3,9 @@
 import os
 import glob
 
-print("Which directory or directories would you like to clear?")
-output_directory = [input("(type all for all directories, separate multiple directories by spaces): ")][0].split(' ')
+print()
+print(" Which directory(ies) from '/examples/' would you like to clear?")
+output_directory = [input(" (type all for all directories, separate multiple directories by spaces): ")][0].split(' ')
 
 os.chdir('./examples/')
 
@@ -15,9 +16,14 @@ for directory in output_directory:
   if os.path.isdir(directory):
     fileNames = glob.glob(f'./{directory}/*.out.*')
     fileNames = fileNames + glob.glob(f'./{directory}/*.pfidb')
-    try:
-      for file in fileNames:
-        os.remove(file)
-      print(f'Cleared directory "{directory}"')
-    except OSError:
-      print(f'Error in clearing directory "{directory}"')
+    if len(fileNames):
+      try:
+        for file in fileNames:
+          os.remove(file)
+        print(f'  + Cleared directory "{directory}"')
+      except OSError:
+        print(f'  X Error in clearing directory "{directory}"')
+    else:
+      print(f'  - No output files in {directory}')
+
+print()
